+++ npc/util/core.py
@@ -27,19 +27,36 @@ def load_json(filename):
     Returns:
         List or dict from `json.loads()`
     """
-    comment_re = re.compile(
-        r'(^)?[^\S\n]*/(?:\*(.*?)\*/[^\S\n]*|/[^\n]*)($)?',
-        re.DOTALL | re.MULTILINE
-    )
+
+    def remove_comments(json_like):
+        """
+        Removes C-style comments from *json_like* and returns the result.
+        """
+        comments_re = re.compile(
+            r'//.*?$|/\*.*?\*/|\'(?:\\.|[^\\\'])*\'|"(?:\\.|[^\\"])*"',
+            re.DOTALL | re.MULTILINE
+        )
+        def replacer(match):
+            s = match.group(0)
+            if s[0] == '/': return ""
+            return s
+        return comments_re.sub(replacer, json_like)
+
+    def remove_trailing_commas(json_like):
+        """
+        Removes trailing commas from *json_like* and returns the result.
+        """
+        trailing_object_commas_re = re.compile(r'(,)\s*}(?=([^"\\]*(\\.|"([^"\\]*\\.)*[^"\\]*"))*[^"]*$)')
+        trailing_array_commas_re = re.compile(r'(,)\s*\](?=([^"\\]*(\\.|"([^"\\]*\\.)*[^"\\]*"))*[^"]*$)')
+        # Fix objects {} first
+        objects_fixed = trailing_object_commas_re.sub("}", json_like)
+        # Now fix arrays/lists [] and return the result
+        return trailing_array_commas_re.sub("]", objects_fixed)
+
     with open(filename) as json_file:
         content = ''.join(json_file.readlines())
-
-        ## Looking for comments
-        match = comment_re.search(content)
-        while match:
-            # single line comment
-            content = content[:match.start()] + content[match.end():]
-            match = comment_re.search(content)
+        content = remove_comments(content) # Remove comments
+        content = remove_trailing_commas(content) # Remove trailing commas
 
         # Return parsed json
         try: